+++ aiarena/frontend-spa/src/_pages/AdminCompetitions/utils/DateValue.tsx
@@ -1,0 +1,5 @@
+import { formatDate } from "./formatDate";
+
+export function DateValue({ value }: { value: string | null | undefined }) {
+  return <>{formatDate(value)}</>;
+}
