+++ aiarena/frontend-spa/src/_pages/AdminCompetitions/utils/formatDate.ts
@@ -1,0 +1,7 @@
+export function formatDate(value: string | null | undefined) {
+  if (!value) {
+    return "—";
+  }
+
+  return new Date(value).toLocaleDateString();
+}
