+++ aiarena/frontend-spa/src/_pages/AdminCompetitions/utils/formatDateTime.ts
@@ -1,0 +1,7 @@
+export function formatDateTime(value: string | null | undefined) {
+  if (!value) {
+    return "—";
+  }
+
+  return new Date(value).toLocaleString();
+}
