+++ aiarena/graphql/mutations.py
@@ -602,6 +602,7 @@ class CheckCompetitionTrophiesInput(CleanedInputType):
     class Meta:
         required_fields = ["competition"]
 
+
 class CheckCompetitionTrophies(CleanedInputMutation):
     status = graphene.Field(
         CompetitionTrophyCheckStatus,
@@ -680,9 +681,7 @@ class CheckCompetitionTrophies(CleanedInputMutation):
                 )
                 for expected in report.missing_trophies
             ],
-            incorrect_trophy_ids=[
-                str(trophy.id) for trophy in report.incorrect_trophies
-            ],
+            incorrect_trophy_ids=[str(trophy.id) for trophy in report.incorrect_trophies],
             incorrect_trophies=[
                 IncorrectCompetitionTrophyType(
                     id=trophy.id,
@@ -761,12 +760,8 @@ class AwardCompetitionTrophies(CleanedInputMutation):
             awards_given=competition.awards_given,
             created_trophy_count=report.created_trophy_count,
             deleted_trophy_count=report.deleted_trophy_count,
-            created_trophy_ids=[
-                str(trophy_id) for trophy_id in report.created_trophy_ids
-            ],
-            deleted_trophy_ids=[
-                str(trophy_id) for trophy_id in report.deleted_trophy_ids
-            ],
+            created_trophy_ids=[str(trophy_id) for trophy_id in report.created_trophy_ids],
+            deleted_trophy_ids=[str(trophy_id) for trophy_id in report.deleted_trophy_ids],
         )
